pyproject.toml
[project]
name = "nanobot-runtime"
version = "0.1.0"
description = "Shared runtime glue for nanobot workspaces: AgentLoop hooks, gateway launcher, LTM plumbing"
requires-python = ">=3.12"

dependencies = [
    "nanobot-ai>=0.1.5.post1",
    "fastmcp>=2.0",
    "loguru>=0.7",
    "fast-bunkai>=0.1.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.30",
]

[tool.uv.sources]
nanobot-ai = { git = "https://github.com/yw0nam/nanobot", branch = "develop" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nanobot_runtime"]

# Workspace launcher console script. Workspaces install nanobot-runtime as
# an editable dep and invoke ``nanobot-launcher`` (or ``python -m
# nanobot_runtime.launcher``) instead of carrying their own entrypoint file.
[project.scripts]
nanobot-launcher = "nanobot_runtime.launcher:main"

# Nanobot's ChannelManager discovers external channel plugins through the
# ``nanobot.channels`` entry_point group. The section key in nanobot.json
# (``channels.desktop_mate``) must match the entry_point name below.
[project.entry-points."nanobot.channels"]
desktop_mate = "nanobot_runtime.services.channels.desktop_mate:DesktopMateChannel"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that require a running ltm-mcp server",
    "e2e: live-infrastructure tests (spawn a real gateway against vLLM/Irodori/LTM)",
]
addopts = "-m 'not integration and not e2e'"
